@@ -1,12 +1,12 @@
 name: image_picker_ios
 description: iOS implementation of the image_picker plugin.
-repository: https://github.com/flutter/plugins/tree/main/packages/image_picker/image_picker_ios
+repository: https://github.com/flutter/packages/tree/main/packages/image_picker/image_picker_ios
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+image_picker%22
-version: 0.8.6+5
+version: 0.8.8+2
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=2.10.0"
+  sdk: ">=2.19.0 <4.0.0"
+  flutter: ">=3.7.0"
 
 flutter:
   plugin:
@@ -19,10 +19,16 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  image_picker_platform_interface: ^2.6.1
+  image_picker_platform_interface: ^2.8.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^5.0.0
-  pigeon: ^3.0.2
+  mockito: 5.4.1
+  pigeon: ^9.2.4
+
+topics:
+  - camera
+  - image-picker
+  - files
+  - file-selection
