@@ -36,12 +36,21 @@
 }
 
 - (void)testScaledImage_ShouldBeCorrectRotation {
-  UIImage *image = [UIImage imageWithData:ImagePickerTestImages.JPGTestData];
-  UIImage *newImage = [FLTImagePickerImageUtil scaledImage:image
-                                                  maxWidth:@3
-                                                 maxHeight:@2
-                                       isMetadataAvailable:YES];
+  NSURL *imageURL =
+      [[NSBundle bundleForClass:[self class]] URLForResource:@"jpgImageWithRightOrientation"
+                                               withExtension:@"jpg"];
+  NSData *imageData = [NSData dataWithContentsOfURL:imageURL];
+  UIImage *image = [UIImage imageWithData:imageData];
+  XCTAssertEqual(image.size.width, 130);
+  XCTAssertEqual(image.size.height, 174);
+  XCTAssertEqual(image.imageOrientation, UIImageOrientationRight);
 
+  UIImage *newImage = [FLTImagePickerImageUtil scaledImage:image
+                                                  maxWidth:@10
+                                                 maxHeight:@10
+                                       isMetadataAvailable:YES];
+  XCTAssertEqual(newImage.size.width, 10);
+  XCTAssertEqual(newImage.size.height, 7);
   XCTAssertEqual(newImage.imageOrientation, UIImageOrientationUp);
 }
 
