@@ -11,7 +11,7 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /**
- * The return hander used for all method calls, which internally adapts the provided result list
+ * The return handler used for all method calls, which internally adapts the provided result list
  * to return either a list or a single element depending on the original call.
  */
 typedef void (^FlutterResultAdapter)(NSArray<NSString *> *_Nullable, FlutterError *_Nullable);
@@ -48,6 +48,9 @@ typedef void (^FlutterResultAdapter)(NSArray<NSString *> *_Nullable, FlutterErro
 
 /** Whether the image should be picked with full metadata (requires gallery permissions) */
 @property(nonatomic, assign) BOOL requestFullMetadata;
+
+/** Whether the picker should include videos in the list*/
+@property(nonatomic, assign) BOOL includeVideo;
 
 @end
 
