@@ -1,9 +1,11 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v3.0.3), do not edit directly.
+// Autogenerated from Pigeon (v9.2.5), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
+
 #import <Foundation/Foundation.h>
+
 @protocol FlutterBinaryMessenger;
 @protocol FlutterMessageCodec;
 @class FlutterError;
@@ -22,12 +24,26 @@ typedef NS_ENUM(NSUInteger, FLTSourceType) {
 };
 
 @class FLTMaxSize;
+@class FLTMediaSelectionOptions;
 @class FLTSourceSpecification;
 
 @interface FLTMaxSize : NSObject
 + (instancetype)makeWithWidth:(nullable NSNumber *)width height:(nullable NSNumber *)height;
 @property(nonatomic, strong, nullable) NSNumber *width;
 @property(nonatomic, strong, nullable) NSNumber *height;
+@end
+
+@interface FLTMediaSelectionOptions : NSObject
+/// `init` unavailable to enforce nonnull fields, see the `make` class method.
+- (instancetype)init NS_UNAVAILABLE;
++ (instancetype)makeWithMaxSize:(FLTMaxSize *)maxSize
+                   imageQuality:(nullable NSNumber *)imageQuality
+            requestFullMetadata:(NSNumber *)requestFullMetadata
+                  allowMultiple:(NSNumber *)allowMultiple;
+@property(nonatomic, strong) FLTMaxSize *maxSize;
+@property(nonatomic, strong, nullable) NSNumber *imageQuality;
+@property(nonatomic, strong) NSNumber *requestFullMetadata;
+@property(nonatomic, strong) NSNumber *allowMultiple;
 @end
 
 @interface FLTSourceSpecification : NSObject
@@ -55,6 +71,10 @@ NSObject<FlutterMessageCodec> *FLTImagePickerApiGetCodec(void);
 - (void)pickVideoWithSource:(FLTSourceSpecification *)source
                 maxDuration:(nullable NSNumber *)maxDurationSeconds
                  completion:(void (^)(NSString *_Nullable, FlutterError *_Nullable))completion;
+/// Selects images and videos and returns their paths.
+- (void)pickMediaWithMediaSelectionOptions:(FLTMediaSelectionOptions *)mediaSelectionOptions
+                                completion:(void (^)(NSArray<NSString *> *_Nullable,
+                                                     FlutterError *_Nullable))completion;
 @end
 
 extern void FLTImagePickerApiSetup(id<FlutterBinaryMessenger> binaryMessenger,
