@@ -1,8 +1,9 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v3.0.3), do not edit directly.
+// Autogenerated from Pigeon (v9.2.5), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
+
 #import "messages.g.h"
 #import <Flutter/Flutter.h>
 
@@ -10,23 +11,13 @@
 #error File requires ARC to be enabled.
 #endif
 
-static NSDictionary<NSString *, id> *wrapResult(id result, FlutterError *error) {
-  NSDictionary *errorDict = (NSDictionary *)[NSNull null];
+static NSArray *wrapResult(id result, FlutterError *error) {
   if (error) {
-    errorDict = @{
-      @"code" : (error.code ? error.code : [NSNull null]),
-      @"message" : (error.message ? error.message : [NSNull null]),
-      @"details" : (error.details ? error.details : [NSNull null]),
-    };
+    return @[
+      error.code ?: [NSNull null], error.message ?: [NSNull null], error.details ?: [NSNull null]
+    ];
   }
-  return @{
-    @"result" : (result ? result : [NSNull null]),
-    @"error" : errorDict,
-  };
-}
-static id GetNullableObject(NSDictionary *dict, id key) {
-  id result = dict[key];
-  return (result == [NSNull null]) ? nil : result;
+  return @[ result ?: [NSNull null] ];
 }
 static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
   id result = array[key];
@@ -34,12 +25,21 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 }
 
 @interface FLTMaxSize ()
-+ (FLTMaxSize *)fromMap:(NSDictionary *)dict;
-- (NSDictionary *)toMap;
++ (FLTMaxSize *)fromList:(NSArray *)list;
++ (nullable FLTMaxSize *)nullableFromList:(NSArray *)list;
+- (NSArray *)toList;
 @end
+
+@interface FLTMediaSelectionOptions ()
++ (FLTMediaSelectionOptions *)fromList:(NSArray *)list;
++ (nullable FLTMediaSelectionOptions *)nullableFromList:(NSArray *)list;
+- (NSArray *)toList;
+@end
+
 @interface FLTSourceSpecification ()
-+ (FLTSourceSpecification *)fromMap:(NSDictionary *)dict;
-- (NSDictionary *)toMap;
++ (FLTSourceSpecification *)fromList:(NSArray *)list;
++ (nullable FLTSourceSpecification *)nullableFromList:(NSArray *)list;
+- (NSArray *)toList;
 @end
 
 @implementation FLTMaxSize
@@ -49,16 +49,56 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
   pigeonResult.height = height;
   return pigeonResult;
 }
-+ (FLTMaxSize *)fromMap:(NSDictionary *)dict {
++ (FLTMaxSize *)fromList:(NSArray *)list {
   FLTMaxSize *pigeonResult = [[FLTMaxSize alloc] init];
-  pigeonResult.width = GetNullableObject(dict, @"width");
-  pigeonResult.height = GetNullableObject(dict, @"height");
+  pigeonResult.width = GetNullableObjectAtIndex(list, 0);
+  pigeonResult.height = GetNullableObjectAtIndex(list, 1);
   return pigeonResult;
 }
-- (NSDictionary *)toMap {
-  return [NSDictionary
-      dictionaryWithObjectsAndKeys:(self.width ? self.width : [NSNull null]), @"width",
-                                   (self.height ? self.height : [NSNull null]), @"height", nil];
++ (nullable FLTMaxSize *)nullableFromList:(NSArray *)list {
+  return (list) ? [FLTMaxSize fromList:list] : nil;
+}
+- (NSArray *)toList {
+  return @[
+    (self.width ?: [NSNull null]),
+    (self.height ?: [NSNull null]),
+  ];
+}
+@end
+
+@implementation FLTMediaSelectionOptions
++ (instancetype)makeWithMaxSize:(FLTMaxSize *)maxSize
+                   imageQuality:(nullable NSNumber *)imageQuality
+            requestFullMetadata:(NSNumber *)requestFullMetadata
+                  allowMultiple:(NSNumber *)allowMultiple {
+  FLTMediaSelectionOptions *pigeonResult = [[FLTMediaSelectionOptions alloc] init];
+  pigeonResult.maxSize = maxSize;
+  pigeonResult.imageQuality = imageQuality;
+  pigeonResult.requestFullMetadata = requestFullMetadata;
+  pigeonResult.allowMultiple = allowMultiple;
+  return pigeonResult;
+}
++ (FLTMediaSelectionOptions *)fromList:(NSArray *)list {
+  FLTMediaSelectionOptions *pigeonResult = [[FLTMediaSelectionOptions alloc] init];
+  pigeonResult.maxSize = [FLTMaxSize nullableFromList:(GetNullableObjectAtIndex(list, 0))];
+  NSAssert(pigeonResult.maxSize != nil, @"");
+  pigeonResult.imageQuality = GetNullableObjectAtIndex(list, 1);
+  pigeonResult.requestFullMetadata = GetNullableObjectAtIndex(list, 2);
+  NSAssert(pigeonResult.requestFullMetadata != nil, @"");
+  pigeonResult.allowMultiple = GetNullableObjectAtIndex(list, 3);
+  NSAssert(pigeonResult.allowMultiple != nil, @"");
+  return pigeonResult;
+}
++ (nullable FLTMediaSelectionOptions *)nullableFromList:(NSArray *)list {
+  return (list) ? [FLTMediaSelectionOptions fromList:list] : nil;
+}
+- (NSArray *)toList {
+  return @[
+    (self.maxSize ? [self.maxSize toList] : [NSNull null]),
+    (self.imageQuality ?: [NSNull null]),
+    (self.requestFullMetadata ?: [NSNull null]),
+    (self.allowMultiple ?: [NSNull null]),
+  ];
 }
 @end
 
@@ -69,15 +109,20 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
   pigeonResult.camera = camera;
   return pigeonResult;
 }
-+ (FLTSourceSpecification *)fromMap:(NSDictionary *)dict {
++ (FLTSourceSpecification *)fromList:(NSArray *)list {
   FLTSourceSpecification *pigeonResult = [[FLTSourceSpecification alloc] init];
-  pigeonResult.type = [GetNullableObject(dict, @"type") integerValue];
-  pigeonResult.camera = [GetNullableObject(dict, @"camera") integerValue];
+  pigeonResult.type = [GetNullableObjectAtIndex(list, 0) integerValue];
+  pigeonResult.camera = [GetNullableObjectAtIndex(list, 1) integerValue];
   return pigeonResult;
 }
-- (NSDictionary *)toMap {
-  return [NSDictionary
-      dictionaryWithObjectsAndKeys:@(self.type), @"type", @(self.camera), @"camera", nil];
++ (nullable FLTSourceSpecification *)nullableFromList:(NSArray *)list {
+  return (list) ? [FLTSourceSpecification fromList:list] : nil;
+}
+- (NSArray *)toList {
+  return @[
+    @(self.type),
+    @(self.camera),
+  ];
 }
 @end
 
@@ -87,11 +132,11 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 - (nullable id)readValueOfType:(UInt8)type {
   switch (type) {
     case 128:
-      return [FLTMaxSize fromMap:[self readValue]];
-
+      return [FLTMaxSize fromList:[self readValue]];
     case 129:
-      return [FLTSourceSpecification fromMap:[self readValue]];
-
+      return [FLTMediaSelectionOptions fromList:[self readValue]];
+    case 130:
+      return [FLTSourceSpecification fromList:[self readValue]];
     default:
       return [super readValueOfType:type];
   }
@@ -104,10 +149,13 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 - (void)writeValue:(id)value {
   if ([value isKindOfClass:[FLTMaxSize class]]) {
     [self writeByte:128];
-    [self writeValue:[value toMap]];
-  } else if ([value isKindOfClass:[FLTSourceSpecification class]]) {
+    [self writeValue:[value toList]];
+  } else if ([value isKindOfClass:[FLTMediaSelectionOptions class]]) {
     [self writeByte:129];
-    [self writeValue:[value toMap]];
+    [self writeValue:[value toList]];
+  } else if ([value isKindOfClass:[FLTSourceSpecification class]]) {
+    [self writeByte:130];
+    [self writeValue:[value toList]];
   } else {
     [super writeValue:value];
   }
@@ -125,9 +173,9 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 }
 @end
 
-NSObject<FlutterMessageCodec> *FLTImagePickerApiGetCodec() {
-  static dispatch_once_t sPred = 0;
+NSObject<FlutterMessageCodec> *FLTImagePickerApiGetCodec(void) {
   static FlutterStandardMessageCodec *sSharedObject = nil;
+  static dispatch_once_t sPred = 0;
   dispatch_once(&sPred, ^{
     FLTImagePickerApiCodecReaderWriter *readerWriter =
         [[FLTImagePickerApiCodecReaderWriter alloc] init];
@@ -214,6 +262,30 @@ void FLTImagePickerApiSetup(id<FlutterBinaryMessenger> binaryMessenger,
                       completion:^(NSString *_Nullable output, FlutterError *_Nullable error) {
                         callback(wrapResult(output, error));
                       }];
+      }];
+    } else {
+      [channel setMessageHandler:nil];
+    }
+  }
+  /// Selects images and videos and returns their paths.
+  {
+    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
+           initWithName:@"dev.flutter.pigeon.ImagePickerApi.pickMedia"
+        binaryMessenger:binaryMessenger
+                  codec:FLTImagePickerApiGetCodec()];
+    if (api) {
+      NSCAssert([api respondsToSelector:@selector(pickMediaWithMediaSelectionOptions:completion:)],
+                @"FLTImagePickerApi api (%@) doesn't respond to "
+                @"@selector(pickMediaWithMediaSelectionOptions:completion:)",
+                api);
+      [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
+        NSArray *args = message;
+        FLTMediaSelectionOptions *arg_mediaSelectionOptions = GetNullableObjectAtIndex(args, 0);
+        [api pickMediaWithMediaSelectionOptions:arg_mediaSelectionOptions
+                                     completion:^(NSArray<NSString *> *_Nullable output,
+                                                  FlutterError *_Nullable error) {
+                                       callback(wrapResult(output, error));
+                                     }];
       }];
     } else {
       [channel setMessageHandler:nil];
