@@ -1,14 +1,14 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v3.0.3), do not edit directly.
+// Autogenerated from Pigeon (v9.2.5), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
-// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name
-// @dart = 2.12
-import 'dart:async';
-import 'dart:typed_data' show Uint8List, Int32List, Int64List, Float64List;
+// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name, unnecessary_import
 
-import 'package:flutter/foundation.dart' show WriteBuffer, ReadBuffer;
+import 'dart:async';
+import 'dart:typed_data' show Float64List, Int32List, Int64List, Uint8List;
+
+import 'package:flutter/foundation.dart' show ReadBuffer, WriteBuffer;
 import 'package:flutter/services.dart';
 
 enum SourceCamera {
@@ -28,20 +28,57 @@ class MaxSize {
   });
 
   double? width;
+
   double? height;
 
   Object encode() {
-    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
-    pigeonMap['width'] = width;
-    pigeonMap['height'] = height;
-    return pigeonMap;
+    return <Object?>[
+      width,
+      height,
+    ];
   }
 
-  static MaxSize decode(Object message) {
-    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
+  static MaxSize decode(Object result) {
+    result as List<Object?>;
     return MaxSize(
-      width: pigeonMap['width'] as double?,
-      height: pigeonMap['height'] as double?,
+      width: result[0] as double?,
+      height: result[1] as double?,
+    );
+  }
+}
+
+class MediaSelectionOptions {
+  MediaSelectionOptions({
+    required this.maxSize,
+    this.imageQuality,
+    required this.requestFullMetadata,
+    required this.allowMultiple,
+  });
+
+  MaxSize maxSize;
+
+  int? imageQuality;
+
+  bool requestFullMetadata;
+
+  bool allowMultiple;
+
+  Object encode() {
+    return <Object?>[
+      maxSize.encode(),
+      imageQuality,
+      requestFullMetadata,
+      allowMultiple,
+    ];
+  }
+
+  static MediaSelectionOptions decode(Object result) {
+    result as List<Object?>;
+    return MediaSelectionOptions(
+      maxSize: MaxSize.decode(result[0]! as List<Object?>),
+      imageQuality: result[1] as int?,
+      requestFullMetadata: result[2]! as bool,
+      allowMultiple: result[3]! as bool,
     );
   }
 }
@@ -53,22 +90,21 @@ class SourceSpecification {
   });
 
   SourceType type;
+
   SourceCamera? camera;
 
   Object encode() {
-    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
-    pigeonMap['type'] = type.index;
-    pigeonMap['camera'] = camera?.index;
-    return pigeonMap;
+    return <Object?>[
+      type.index,
+      camera?.index,
+    ];
   }
 
-  static SourceSpecification decode(Object message) {
-    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
+  static SourceSpecification decode(Object result) {
+    result as List<Object?>;
     return SourceSpecification(
-      type: SourceType.values[pigeonMap['type']! as int],
-      camera: pigeonMap['camera'] != null
-          ? SourceCamera.values[pigeonMap['camera']! as int]
-          : null,
+      type: SourceType.values[result[0]! as int],
+      camera: result[1] != null ? SourceCamera.values[result[1]! as int] : null,
     );
   }
 }
@@ -80,8 +116,11 @@ class _ImagePickerApiCodec extends StandardMessageCodec {
     if (value is MaxSize) {
       buffer.putUint8(128);
       writeValue(buffer, value.encode());
-    } else if (value is SourceSpecification) {
+    } else if (value is MediaSelectionOptions) {
       buffer.putUint8(129);
+      writeValue(buffer, value.encode());
+    } else if (value is SourceSpecification) {
+      buffer.putUint8(130);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -93,10 +132,10 @@ class _ImagePickerApiCodec extends StandardMessageCodec {
     switch (type) {
       case 128:
         return MaxSize.decode(readValue(buffer)!);
-
       case 129:
+        return MediaSelectionOptions.decode(readValue(buffer)!);
+      case 130:
         return SourceSpecification.decode(readValue(buffer)!);
-
       default:
         return super.readValueOfType(type, buffer);
     }
@@ -109,7 +148,6 @@ class ImagePickerApi {
   /// BinaryMessenger will be used which routes to the host platform.
   ImagePickerApi({BinaryMessenger? binaryMessenger})
       : _binaryMessenger = binaryMessenger;
-
   final BinaryMessenger? _binaryMessenger;
 
   static const MessageCodec<Object?> codec = _ImagePickerApiCodec();
@@ -119,53 +157,54 @@ class ImagePickerApi {
     final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
         'dev.flutter.pigeon.ImagePickerApi.pickImage', codec,
         binaryMessenger: _binaryMessenger);
-    final Map<Object?, Object?>? replyMap = await channel.send(<Object?>[
+    final List<Object?>? replyList = await channel.send(<Object?>[
       arg_source,
       arg_maxSize,
       arg_imageQuality,
       arg_requestFullMetadata
-    ]) as Map<Object?, Object?>?;
-    if (replyMap == null) {
+    ]) as List<Object?>?;
+    if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
         message: 'Unable to establish connection on channel.',
       );
-    } else if (replyMap['error'] != null) {
-      final Map<Object?, Object?> error =
-          (replyMap['error'] as Map<Object?, Object?>?)!;
+    } else if (replyList.length > 1) {
       throw PlatformException(
-        code: (error['code'] as String?)!,
-        message: error['message'] as String?,
-        details: error['details'],
+        code: replyList[0]! as String,
+        message: replyList[1] as String?,
+        details: replyList[2],
       );
     } else {
-      return (replyMap['result'] as String?);
+      return (replyList[0] as String?);
     }
   }
 
-  Future<List<String?>?> pickMultiImage(MaxSize arg_maxSize,
+  Future<List<String?>> pickMultiImage(MaxSize arg_maxSize,
       int? arg_imageQuality, bool arg_requestFullMetadata) async {
     final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
         'dev.flutter.pigeon.ImagePickerApi.pickMultiImage', codec,
         binaryMessenger: _binaryMessenger);
-    final Map<Object?, Object?>? replyMap = await channel.send(
+    final List<Object?>? replyList = await channel.send(
             <Object?>[arg_maxSize, arg_imageQuality, arg_requestFullMetadata])
-        as Map<Object?, Object?>?;
-    if (replyMap == null) {
+        as List<Object?>?;
+    if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
         message: 'Unable to establish connection on channel.',
       );
-    } else if (replyMap['error'] != null) {
-      final Map<Object?, Object?> error =
-          (replyMap['error'] as Map<Object?, Object?>?)!;
+    } else if (replyList.length > 1) {
       throw PlatformException(
-        code: (error['code'] as String?)!,
-        message: error['message'] as String?,
-        details: error['details'],
+        code: replyList[0]! as String,
+        message: replyList[1] as String?,
+        details: replyList[2],
+      );
+    } else if (replyList[0] == null) {
+      throw PlatformException(
+        code: 'null-error',
+        message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (replyMap['result'] as List<Object?>?)?.cast<String?>();
+      return (replyList[0] as List<Object?>?)!.cast<String?>();
     }
   }
 
@@ -174,24 +213,50 @@ class ImagePickerApi {
     final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
         'dev.flutter.pigeon.ImagePickerApi.pickVideo', codec,
         binaryMessenger: _binaryMessenger);
-    final Map<Object?, Object?>? replyMap =
-        await channel.send(<Object?>[arg_source, arg_maxDurationSeconds])
-            as Map<Object?, Object?>?;
-    if (replyMap == null) {
+    final List<Object?>? replyList = await channel
+        .send(<Object?>[arg_source, arg_maxDurationSeconds]) as List<Object?>?;
+    if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
         message: 'Unable to establish connection on channel.',
       );
-    } else if (replyMap['error'] != null) {
-      final Map<Object?, Object?> error =
-          (replyMap['error'] as Map<Object?, Object?>?)!;
+    } else if (replyList.length > 1) {
       throw PlatformException(
-        code: (error['code'] as String?)!,
-        message: error['message'] as String?,
-        details: error['details'],
+        code: replyList[0]! as String,
+        message: replyList[1] as String?,
+        details: replyList[2],
       );
     } else {
-      return (replyMap['result'] as String?);
+      return (replyList[0] as String?);
+    }
+  }
+
+  /// Selects images and videos and returns their paths.
+  Future<List<String?>> pickMedia(
+      MediaSelectionOptions arg_mediaSelectionOptions) async {
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.ImagePickerApi.pickMedia', codec,
+        binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel
+        .send(<Object?>[arg_mediaSelectionOptions]) as List<Object?>?;
+    if (replyList == null) {
+      throw PlatformException(
+        code: 'channel-error',
+        message: 'Unable to establish connection on channel.',
+      );
+    } else if (replyList.length > 1) {
+      throw PlatformException(
+        code: replyList[0]! as String,
+        message: replyList[1] as String?,
+        details: replyList[2],
+      );
+    } else if (replyList[0] == null) {
+      throw PlatformException(
+        code: 'null-error',
+        message: 'Host platform returned null value for non-null return value.',
+      );
+    } else {
+      return (replyList[0] as List<Object?>?)!.cast<String?>();
     }
   }
 }
