@@ -1,20 +1,16 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v3.0.3), do not edit directly.
+// Autogenerated from Pigeon (v9.2.5), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
-// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis
+// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, unnecessary_import
 // ignore_for_file: avoid_relative_lib_imports
-// @dart = 2.12
-import 'dart:typed_data' show Uint8List, Int32List, Int64List, Float64List;
-
-// TODO(a14n): remove this import once Flutter 3.1 or later reaches stable (including flutter/flutter#106316)
-// ignore: unnecessary_import
-import 'package:flutter/foundation.dart' show WriteBuffer, ReadBuffer;
+import 'dart:async';
+import 'dart:typed_data' show Float64List, Int32List, Int64List, Uint8List;
+import 'package:flutter/foundation.dart' show ReadBuffer, WriteBuffer;
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-// Manually changed due to https://github.com/flutter/flutter/issues/97744
 import 'package:image_picker_ios/src/messages.g.dart';
 
 class _TestHostImagePickerApiCodec extends StandardMessageCodec {
@@ -24,8 +20,11 @@ class _TestHostImagePickerApiCodec extends StandardMessageCodec {
     if (value is MaxSize) {
       buffer.putUint8(128);
       writeValue(buffer, value.encode());
-    } else if (value is SourceSpecification) {
+    } else if (value is MediaSelectionOptions) {
       buffer.putUint8(129);
+      writeValue(buffer, value.encode());
+    } else if (value is SourceSpecification) {
+      buffer.putUint8(130);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -37,10 +36,10 @@ class _TestHostImagePickerApiCodec extends StandardMessageCodec {
     switch (type) {
       case 128:
         return MaxSize.decode(readValue(buffer)!);
-
       case 129:
+        return MediaSelectionOptions.decode(readValue(buffer)!);
+      case 130:
         return SourceSpecification.decode(readValue(buffer)!);
-
       default:
         return super.readValueOfType(type, buffer);
     }
@@ -48,14 +47,22 @@ class _TestHostImagePickerApiCodec extends StandardMessageCodec {
 }
 
 abstract class TestHostImagePickerApi {
+  static TestDefaultBinaryMessengerBinding? get _testBinaryMessengerBinding =>
+      TestDefaultBinaryMessengerBinding.instance;
   static const MessageCodec<Object?> codec = _TestHostImagePickerApiCodec();
 
   Future<String?> pickImage(SourceSpecification source, MaxSize maxSize,
       int? imageQuality, bool requestFullMetadata);
-  Future<List<String?>?> pickMultiImage(
+
+  Future<List<String?>> pickMultiImage(
       MaxSize maxSize, int? imageQuality, bool requestFullMetadata);
+
   Future<String?> pickVideo(
       SourceSpecification source, int? maxDurationSeconds);
+
+  /// Selects images and videos and returns their paths.
+  Future<List<String?>> pickMedia(MediaSelectionOptions mediaSelectionOptions);
+
   static void setup(TestHostImagePickerApi? api,
       {BinaryMessenger? binaryMessenger}) {
     {
@@ -63,9 +70,12 @@ abstract class TestHostImagePickerApi {
           'dev.flutter.pigeon.ImagePickerApi.pickImage', codec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMockMessageHandler(null);
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel, null);
       } else {
-        channel.setMockMessageHandler((Object? message) async {
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel,
+                (Object? message) async {
           assert(message != null,
               'Argument for dev.flutter.pigeon.ImagePickerApi.pickImage was null.');
           final List<Object?> args = (message as List<Object?>?)!;
@@ -82,7 +92,7 @@ abstract class TestHostImagePickerApi {
               'Argument for dev.flutter.pigeon.ImagePickerApi.pickImage was null, expected non-null bool.');
           final String? output = await api.pickImage(arg_source!, arg_maxSize!,
               arg_imageQuality, arg_requestFullMetadata!);
-          return <Object?, Object?>{'result': output};
+          return <Object?>[output];
         });
       }
     }
@@ -91,9 +101,12 @@ abstract class TestHostImagePickerApi {
           'dev.flutter.pigeon.ImagePickerApi.pickMultiImage', codec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMockMessageHandler(null);
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel, null);
       } else {
-        channel.setMockMessageHandler((Object? message) async {
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel,
+                (Object? message) async {
           assert(message != null,
               'Argument for dev.flutter.pigeon.ImagePickerApi.pickMultiImage was null.');
           final List<Object?> args = (message as List<Object?>?)!;
@@ -104,9 +117,9 @@ abstract class TestHostImagePickerApi {
           final bool? arg_requestFullMetadata = (args[2] as bool?);
           assert(arg_requestFullMetadata != null,
               'Argument for dev.flutter.pigeon.ImagePickerApi.pickMultiImage was null, expected non-null bool.');
-          final List<String?>? output = await api.pickMultiImage(
+          final List<String?> output = await api.pickMultiImage(
               arg_maxSize!, arg_imageQuality, arg_requestFullMetadata!);
-          return <Object?, Object?>{'result': output};
+          return <Object?>[output];
         });
       }
     }
@@ -115,9 +128,12 @@ abstract class TestHostImagePickerApi {
           'dev.flutter.pigeon.ImagePickerApi.pickVideo', codec,
           binaryMessenger: binaryMessenger);
       if (api == null) {
-        channel.setMockMessageHandler(null);
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel, null);
       } else {
-        channel.setMockMessageHandler((Object? message) async {
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel,
+                (Object? message) async {
           assert(message != null,
               'Argument for dev.flutter.pigeon.ImagePickerApi.pickVideo was null.');
           final List<Object?> args = (message as List<Object?>?)!;
@@ -128,7 +144,31 @@ abstract class TestHostImagePickerApi {
           final int? arg_maxDurationSeconds = (args[1] as int?);
           final String? output =
               await api.pickVideo(arg_source!, arg_maxDurationSeconds);
-          return <Object?, Object?>{'result': output};
+          return <Object?>[output];
+        });
+      }
+    }
+    {
+      final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.ImagePickerApi.pickMedia', codec,
+          binaryMessenger: binaryMessenger);
+      if (api == null) {
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel, null);
+      } else {
+        _testBinaryMessengerBinding!.defaultBinaryMessenger
+            .setMockDecodedMessageHandler<Object?>(channel,
+                (Object? message) async {
+          assert(message != null,
+              'Argument for dev.flutter.pigeon.ImagePickerApi.pickMedia was null.');
+          final List<Object?> args = (message as List<Object?>?)!;
+          final MediaSelectionOptions? arg_mediaSelectionOptions =
+              (args[0] as MediaSelectionOptions?);
+          assert(arg_mediaSelectionOptions != null,
+              'Argument for dev.flutter.pigeon.ImagePickerApi.pickMedia was null, expected non-null MediaSelectionOptions.');
+          final List<String?> output =
+              await api.pickMedia(arg_mediaSelectionOptions!);
+          return <Object?>[output];
         });
       }
     }
